==> junesoo/1.Interpolation.ipynb <==
# Interpolation한 데이터를 합친 후 one-hot encoding, shift, 날짜 데이터 추가하는 페이지

import os
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib as mpl
import matplotlib.pyplot as plt
%matplotlib inline

TRAIN_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN/'
TRAIN_AWS_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN_AWS/'
META_AWS_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/META/awsmap.csv'
META_PM_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/META/pmmap.csv'

LINEAR_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN_AWS_LINEAR/'
POLY_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN_AWS_POLYNOMIAL_2/'

TRAIN_MERGE_LINEAR_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/MERGE_LINEAR/'
TRAIN_MERGE_POLYNOMIAL_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/MERGE_POLYNOMIAL/'

TRAIN_LINEAR_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN_LINEAR/'
TRAIN_POLY_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN_POLYNOMIAL/'

mpl.rcParams['figure.figsize'] = (20,10)
mpl.rcParams['axes.grid'] = False

# shift + Merge

final_merged_df = pd.DataFrame()
for linear in os.listdir(TRAIN_MERGE_LINEAR_PATH):
    df = pd.read_csv(TRAIN_MERGE_LINEAR_PATH+linear)
    df.drop(['year_x','place_y','year_y'],inplace=True,axis=1)
    df.rename(columns={'place_x':'place'},inplace=True)

    for i in range(1,73,1): # (48,120,1)
        df['shift'+str(i)] = df['PM2.5'].shift(-i)
    final_merged_df = pd.concat([final_merged_df,df])

# final_merged_df.dropna(axis=0, inplace=True)

final_merged_df['date'] = pd.to_datetime(final_merged_df['date'], format='%Y-%m-%d %H:%M')
final_merged_df['month'] = final_merged_df['date'].dt.month
final_merged_df['day'] = final_merged_df['date'].dt.day
final_merged_df['hour'] = final_merged_df['date'].dt.hour

final_merged_df = pd.get_dummies(data = final_merged_df, columns=['place','month', 'day', 'hour'])
final_merged_df.to_csv(TRAIN_MERGE_LINEAR_PATH+'final_merged.csv',index=False)

final_merged_df.shape

# shift + Merge

final_merged_df = pd.DataFrame()

for poly in os.listdir(TRAIN_MERGE_POLYNOMIAL_PATH):
    df = pd.read_csv(TRAIN_MERGE_POLYNOMIAL_PATH+poly)
    df.drop(['year_x','place_y','year_y'],inplace=True,axis=1)
    df.rename(columns={'place_x':'place'},inplace=True)

    for i in range(1,73,1): # (48,120,1)
        df['shift'+str(i)] = df['PM2.5'].shift(-i)

    final_merged_df = pd.concat([final_merged_df,df])

# final_merged_df.dropna(axis=0, inplace=True)

final_merged_df['date'] = pd.to_datetime(final_merged_df['date'], format='%Y-%m-%d %H:%M')
final_merged_df['month'] = final_merged_df['date'].dt.month
final_merged_df['day'] = final_merged_df['date'].dt.day
final_merged_df['hour'] = final_merged_df['date'].dt.hour

final_merged_df = pd.get_dummies(data = final_merged_df, columns=['place','month', 'day', 'hour'])
final_merged_df.to_csv(TRAIN_MERGE_POLYNOMIAL_PATH+'final_merged.csv',index=False)

final_merged_df.shape

place_eng = {'place_공주':	'',
            'place_노은동':	'',
            'place_논산':	'',
            'place_대천2동':	'',
            'place_독곶리':	'',
            'place_동문동':	'',
            'place_모종동':	'',
            'place_문창동':	'',
            'place_성성동':	'',
            'place_신방동':	'',
            'place_신흥동':	'',
            'place_아름동':	'',
            'place_예산군':	'',
            'place_읍내동':	'',
            'place_이원면':	'',
            'place_정림동':	'',
            'place_홍성읍':''}
