==> junesoo/3.train_data_generate.ipynb <==
import os
import pandas as pd
import numpy as np

TRAIN_MERGE_LINEAR_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/MERGE_LINEAR/'
TRAIN_MERGE_POLYNOMIAL_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/MERGE_POLYNOMIAL/'

TRAIN_LINEAR_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN_LINEAR/'
TRAIN_POLY_PATH = 'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/data/TRAIN_POLYNOMIAL/'



place_lst = [
    '공주',
    '노은동',
    '논산',
    '대천2동',
    '독곶리',
    '동문동',
    '모종동',
    '문창동',
    '성성동',
    '신방동',
    '신흥동',
    '아름동',
    '예산군',
    '읍내동',
    '이원면',
    '정림동',
    '홍성읍'
    ]


df = pd.read_csv(TRAIN_MERGE_LINEAR_PATH+'final_merged.csv')
df.drop(['date'], axis=1, inplace=True) # ,'PM2.5'
df_all = df.copy()
df_all.shape

for place in place_lst:
    globals()['df_{}'.format(place)] = df_all[df_all['place_{}'.format(place)]==1]

timestep_input = 48
timestep_output = 72
timestep = timestep_input + timestep_output

Encoder_X_lst = []


for place in place_lst:

    df = globals()['df_{}'.format(place)]
    encoder_x = []

    for i in range( 0, len(df) - timestep_output - timestep_input): 
        encoder_x.extend([df.iloc[i : i + timestep_input, :].values])

Encoder_X = np.array(encoder_x)

Decoder_X_lst = []


for place in place_lst:

    df = globals()['df_{}'.format(place)]
    decoder_x = []

    for i in range( 0, len(df) - timestep_output - timestep_input ):
        decoder_x.extend([df.iloc[i + timestep_input : i + timestep_input + timestep_output, 6:].values])
        
Decoder_X = np.array(decoder_x)

Decoder_Y_lst = []


for place in place_lst:

    df = globals()['df_{}'.format(place)]
    decoder_y = []

    for i in range( 0, len(df) - timestep_output - timestep_input ):
        decoder_y.extend([df.iloc[i + timestep_input : i + timestep_input + timestep_output, 0].values.T])
        
Decoder_Y = np.array(decoder_y)

import joblib

LINEAR = 'linear_'
POLYNOMIAL = 'polynomial_'

joblib.dump(Encoder_X,'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/'+LINEAR+'encoder_x.pkl')
joblib.dump(Decoder_Y,'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/'+LINEAR+'y.pkl')
joblib.dump(Decoder_X,'C:/Users/js/Desktop/AISPRAK/사회문제 해결형/'+LINEAR+'decoder_x.pkl')
